Pass trade events using lock free spsc. Draw candlesticks and automatically rescale

+++ include/2fish/models/book_snapshot.h
@@ -5,7 +5,7 @@
 
 namespace market {
 	struct BookSnapshot {
-		std::array<long double, 101> bids_{};
-		std::array<long double, 101> asks_{};
+		std::array<double, 101> bids_{};
+		std::array<double, 101> asks_{};
 	};
 }
