+++ include/2fish/order_book/order_book.h
@@ -18,14 +18,14 @@ namespace market {
 		// TODO: perhaps remove this? we are only using this for unit tests rn
 		uint64_t getSize(Side side, int price);
 
-		const std::array<long double, 101>& getBids() const { return bids_; }
-		const std::array<long double, 101>& getAsks() const { return asks_; }
+		const std::array<double, 101>& getBids() const { return bids_; }
+		const std::array<double, 101>& getAsks() const { return asks_; }
 
 	private:
 		// Instead of traditional tree based approaches for the order book, i use a simple
 		// array because the only possible price levels are between $0 and $1 (100 cents)
-		std::array<long double, 101> bids_{};
-		std::array<long double, 101> asks_{};
+		std::array<double, 101> bids_{};
+		std::array<double, 101> asks_{};
 
 		int best_bid_{};
 		int best_ask_{};
